Major enhancements of the L2 decode:
- Before the function was panicing if it decodes multiple oligos with
the same address, which is wrong. Now the code collects all different
data blocks from all the oligos with the same address and tries all
the combinations of them to see if any of them can be verified with
the erasure codes. This make the code slower, but more accurate.

- Run the decoding in parallel, creating as many threads as cores the
computer has. Should improve the performance loss from the previous
point somewhat.